DexPoolStream
